+++ src/components/CardForm/CardExpirationInput/CardExpirationInput.tsx
@@ -1,10 +1,11 @@
 import { useState, useMemo } from 'react';
-import TitleContainer from './TitleContainer';
-import InputField from './InputField';
-import Input from './Input';
 
-import { CARD_EXPIRATION } from '../constants/Condition';
-import { ERROR_MESSAGE } from '../constants/Message';
+import TitleContainer from '../../common/TitleContainer/TitleContainer';
+import InputField from '../../common/InputField/InputField';
+import Input from '../../common/Input/Input';
+
+import { CARD_EXPIRATION } from '../../../constants/Condition';
+import { ERROR_MESSAGE } from '../../../constants/Message';
 
 interface CardExpirationInputProps {
   setMonth: React.Dispatch<React.SetStateAction<string>>;
