+++ src/components/CardForm/CardNumberInput/CardNumberInput.stories.tsx
@@ -1,5 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import CardNumberInput from '../components/CardNumberInput';
+
+import CardNumberInput from './CardNumberInput';
 
 const meta = {
   title: 'CardNumberInput',
