+++ src/components/common/Input/Input.stories.tsx
@@ -1,5 +1,6 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import Input from '../components/Input';
+
+import Input from './Input';
 
 const meta = {
   title: 'Input',
